requirements.txt
firebase-admin==4.5.1
requests==2.25.1
google-cloud-firestore==2.1.0
Flask==1.1.1
google-cloud-pubsub==2.4.1